requirements.txt
absl-py==0.9.0
astor==0.8.0
certifi==2020.6.20
cycler==0.10.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.27.2
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
opt-einsum==3.1.0
protobuf==3.12.3
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tornado==6.0.4
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1

tensorflow-gpu~=1.15.0
numpy~=1.19.1
scipy~=1.5.0
matplotlib~=3.2.2